+++ backend/ttabong/src/main/java/com/ttabong/dto/recruit/responseDto/org/ReadMyRecruitsResponseDto.java
@@ -2,6 +2,7 @@ package com.ttabong.dto.recruit.responseDto.org;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -51,8 +52,8 @@ public class ReadMyRecruitsResponseDto {
         private Integer maxVolunteer;
         private Integer participateVolCount;
         private LocalDate activityDate;
-        private Double activityStart;
-        private Double activityEnd;
+        private BigDecimal activityStart;
+        private BigDecimal activityEnd;
         private LocalDateTime deadline;
         private LocalDateTime createdAt;
     }
